int to long

+++ src/main/java/com/standard/domain/Marca.java
@@ -1,7 +1,6 @@
 package com.standard.domain;
 
 import com.standard.enums.StatusEnum;
-import com.standard.enums.StatusEnum;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -10,7 +9,7 @@ public @Data class Marca implements Serializable {
 
 	private static final long serialVersionUID = -6612762288260227887L;
 
-	private Integer codigo;
+	private Long codigo;
 	private String nome;
 	private String descricao;
 	private StatusEnum status;
