+++ src/main/java/com/standard/domain/Produto.java
@@ -3,9 +3,6 @@ package com.standard.domain;
 import com.standard.enums.FlagSiteEnum;
 import com.standard.enums.StatusEnum;
 import com.standard.util.Constants;
-import com.standard.enums.FlagSiteEnum;
-import com.standard.enums.StatusEnum;
-import com.standard.util.Constants;
 import lombok.Data;
 import org.springframework.format.annotation.NumberFormat;
 import org.springframework.format.annotation.NumberFormat.Style;
@@ -18,7 +15,7 @@ import java.util.List;
 public @Data class Produto implements Serializable {
 
 	private static final long serialVersionUID = -6612762288260227887L;
-	private Integer codigo;
+	private Long codigo;
 	private String barCode;
 	private String nome;
 	private StatusEnum status;
