+++ src/main/java/com/standard/domain/ProdutoHasItensTipoMedida.java
@@ -9,7 +9,7 @@ public @Data class ProdutoHasItensTipoMedida implements Serializable {
 
 	private static final long serialVersionUID = -6612762288260227887L;
 
-	private Integer codigo;
+	private Long codigo;
 	private List<Dominio> dominios;
 	private Integer quantidade;
 	private Double  valorUnitario;
