+++ src/main/java/com/standard/service/FormaDePagamento/FormaDePagamentoService.java
@@ -1,19 +1,19 @@
 package com.standard.service.FormaDePagamento;
 
-import java.util.List;
-
 import com.standard.domain.FormasDePagamento;
+
+import java.util.List;
 
 public interface FormaDePagamentoService {
 
 	FormasDePagamento incluir(FormasDePagamento objct);
 
-	FormasDePagamento alterar(Integer codigo, FormasDePagamento objct);
+	FormasDePagamento alterar(Long codigo, FormasDePagamento objct);
 	
-	void excluir(Integer codigo);
+	void excluir(Long codigo);
 	
 	List<FormasDePagamento> consultar();
 	
-	FormasDePagamento consultarByCodigo(Integer codigo);
+	FormasDePagamento consultarByCodigo(Long codigo);
 	
 }
